+++ bld/vi/os2/os2sys.c
@@ -300,7 +300,7 @@ bool KeyboardHit( void )
     return( BIOSKeyboardHit() );
 }
 
-void MyVioShowBuf( unsigned offset, unsigned nchars )
+void MyVioShowBuf( size_t offset, unsigned nchars )
 {
     BIOSUpdateScreen( offset, nchars );
 }
